% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL94171-package.R
\docType{package}
\name{PL94171-package}
\alias{PL94171}
\alias{PL94171-package}
\title{PL94171: Tabulate P.L. 94-171 Redistricting Data Summary Files}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tools to process legacy format summary redistricting data files
    produced by the United States Census Bureau pursuant to P.L. 94-171. These
    files are generally available earlier but are difficult to work with as-is.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corymccartan.github.io/PL94171/}
  \item \url{https://github.com/CoryMcCartan/PL94171/}
  \item Report bugs at \url{https://github.com/CoryMcCartan/PL94171/issues}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu}

Authors:
\itemize{
  \item Christopher T. Kenny \email{christopherkenny@fas.harvard.edu}
}

}
\keyword{internal}
