% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_ex.R
\name{pl_ex}
\alias{pl_ex}
\title{PL Example File}
\format{
list of tibbles containing the four PL files.
\describe{
\item{\code{00001}}{Tables P1 and P2}
\item{\code{00002}}{Tables P3, P4, and H1}
\item{\code{00003}}{Table P5}
\item{\code{geo}}{geographic header file}
}
}
\usage{
data("pl_ex")
}
\description{
This data contains a subset of the 2020 prototype PL data
}
\examples{
data(pl_ex)
}
\concept{data}
