% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_get_vtd.R
\name{pl_get_vtd}
\alias{pl_get_vtd}
\title{Download 2020 Voting District Shapefiles}
\usage{
pl_get_vtd(abbr, cache_to = NULL, refresh = FALSE)
}
\arguments{
\item{abbr}{Geography to download data for. See details for full list.}

\item{cache_to}{the file name, if any, to cache the results to (as an RDS).
If a file exists and \code{refresh=FALSE}, will read from this file.}

\item{refresh}{if \code{TRUE}, force a re-download of the data.}
}
\value{
An \code{sf} object containing the VTDs.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A (likely temporary) function to download TIGER shapefiles for 2020 voting
tabulation districts (VTDs).
}
\examples{
\donttest{
shp <- pl_get_vtd("RI")
}

}
\concept{basic}
