% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl_read.R
\name{pl_url}
\alias{pl_url}
\title{Get the URL for PL files for a particular state and year}
\usage{
pl_url(abbr, year = 2010)
}
\arguments{
\item{abbr}{The state to download the PL files for}

\item{year}{The year of PL file to download. Supported years: 1990, 2000, 2010,
2020. 2000 files are in a different format.
Earlier years available on tape or CD-ROM only.}
}
\value{
a character vector containing the URL to a ZIP containing the PL files.
}
\description{
Get the URL for PL files for a particular state and year
}
\examples{
pl_url("RI", 2010)

}
\concept{basic}
