% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_standard_pl}
\alias{select_standard_pl}
\title{Select the Standard Redistricting Columns}
\usage{
select_standard_pl(pl, clean_names = TRUE)
}
\arguments{
\item{pl}{A list of PL tables, as read in by \code{\link[=read_pl]{read_pl()}}}

\item{clean_names}{whether to clean names}
}
\value{
tibble
}
\description{
Selects the standard set of basic population groups and vap groups. Optionally
renames them from the PXXXYYYY naming convention (where XXX is the table and
YYYY is the variable) to more human readable names. pop_* is the total population,
from tables 1 and 2, while vap_* is the 18+ population (voting age population).
}
\details{
If clean names, then the variables extracted are as follows:
\itemize{
\item \verb{\\*_hisp}: Hispanic or Latino (of any race)
\item \verb{\\*_white}: White alone, not Hispanic or Latino
\item \verb{\\*_black}: Black or African American alone, not Hispanic or Latino
\item \verb{\\*_aian}: American Indian and Alaska Native alone, not Hispanic or Latino
\item \verb{\\*_asian}: Asian alone, not Hispanic or Latino
\item \verb{\\*_nhpi}: Native Hawaiian and Other Pacific Islander alone, not Hispanic or Latino
\item \verb{\\*_other}: Some Other Race alone, not Hispanic or Latino
\item \verb{\\*_two}: Population of two or more races, not Hispanic or Latino
}

where \* is \code{pop} or \code{vap}.
}
