% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_pl}
\alias{subset_pl}
\title{Subset ot a Summary Level}
\usage{
subset_pl(pl, sumlev = "750")
}
\arguments{
\item{pl}{A list of PL tables, as read in by \code{\link[=read_pl]{read_pl()}}}

\item{sumlev}{the summary level to filter to. A 3 character SUMLEV code. Default is '750' for blocks.}
}
\value{
tibble
}
\description{
This subsets a pl table to a desired summary level. Typical choices include:
\itemize{
\item '750': block
\item '740': block group
\item '630': voting district
\item '050': county
}
}
\details{
All summary levels are listed in \link{pl_geog_levels}.
}
