% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sumlev}
\alias{sumlev}
\title{List of Summary Levels and Official Descriptions}
\format{
a tibble with two columns:
\describe{
\item{SUMLEV}{The three character summary level code}
\item{SUMLEV_description}{The summary level description}
}
}
\usage{
data('sumlev')
}
\description{
This dataset is tibble version of the descriptions of (potentially) available
summary levels within the P.L. 94-171 data, as described in the 2018 Redistricting
Data Prototype (Public Law 94-171) Summary File documentation.
}
\examples{
data('sumlev')
print(sumlev)
}
