% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pl.R
\name{read_pl}
\alias{read_pl}
\title{Read a set of PL Files}
\usage{
read_pl(path, ...)
}
\arguments{
\item{path}{a path to a folder containing PL files. Can also be path or a URL
for a ZIP file, which will be downloaded and unzipped.}

\item{...}{passed on to \code{\link[vroom:vroom]{vroom::vroom()}}}
}
\value{
A list of \link{tibble}s containing the four PL files.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
PL files come in one of four types and are pipe-delimited with no header row.
This function speedily reads in the files and assigns the appropriate column
names and types.
}
